Rename pkg/adapter/testing to pkg/adapter/test to avoid name conflict.
