+++ adapter/denyChecker/BUILD
@@ -20,5 +20,5 @@ go_test(
     size = "small",
     srcs = ["denyChecker_test.go"],
     library = ":go_default_library",
-    deps = ["//pkg/adapter/testing:go_default_library"],
+    deps = ["//pkg/adapter/test:go_default_library"],
 )
