+++ adapter/genericListChecker/BUILD
@@ -18,5 +18,5 @@ go_test(
     size = "small",
     srcs = ["genericListChecker_test.go"],
     library = ":go_default_library",
-    deps = ["//pkg/adapter/testing:go_default_library"],
+    deps = ["//pkg/adapter/test:go_default_library"],
 )
