+++ adapter/memQuota/BUILD
@@ -22,5 +22,5 @@ go_test(
         "rollingWindow_test.go",
     ],
     library = ":go_default_library",
-    deps = ["//pkg/adapter/testing:go_default_library"],
+    deps = ["//pkg/adapter/test:go_default_library"],
 )
