+++ adapter/stdioLogger/BUILD
@@ -20,7 +20,7 @@ go_test(
     srcs = ["stdioLogger_test.go"],
     library = ":go_default_library",
     deps = [
-        "//pkg/adapter/testing:go_default_library",
+        "//pkg/adapter/test:go_default_library",
         "@com_github_google_go_genproto//googleapis/rpc/status:go_default_library",
     ],
 )
