+++ pkg/adapter/test/BUILD
